requirements.txt
# Secure PR Guard - Docker Compatible Dependencies

# Core AI & API
openai>=1.0.0,<2.0.0
requests>=2.28.0,<3.0.0
python-dotenv>=0.19.0,<2.0.0

# Data Processing
pandas>=1.3.0,<3.0.0
numpy>=1.21.0,<2.0.0

# Visualization
plotly>=5.0.0,<6.0.0

# Configuration
pyyaml>=6.0.0,<7.0.0
click>=8.0.0,<9.0.0

# OpenTelemetry (Essential only)
opentelemetry-api>=1.20.0,<2.0.0
opentelemetry-sdk>=1.20.0,<2.0.0
opentelemetry-exporter-otlp>=1.20.0,<2.0.0

# System
psutil>=5.8.0,<6.0.0

# Enterprise Features
slack-sdk>=3.0.0,<4.0.0

# Testing (minimal)
pytest>=7.0.0,<9.0.0
pytest-cov>=4.0.0,<5.0.0

# Development
typing-extensions>=4.0.0
